Modify Comarison page     - create feature folder in store     - created a separate folder for each component     - remove default export     - remove anonymous functions     - add to each component own styles and interface
