+++ src/pages/comparison/comparison.scss
@@ -36,40 +36,3 @@ span {
     font-size: 16px;
   }
 }
-
-.pet-block {
-  display: flex;
-  align-items: center;
-
-  .pet-image {
-    width: 35px;
-    height: 35px;
-    border-radius: 50px;
-    margin-right: 15px;
-  }
-
-  .pet-name {
-    font-size: 15px;
-    width: 150px;
-    line-height: 16px;
-  }
-}
-
-.item-compare {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin: 10px 0;
-  border-radius: 50px;
-  padding: 3px;
-
-  &:hover {
-    cursor: pointer;
-    background-color: rgb(243, 240, 240);
-  }
-}
-
-.link-style {
-  color: black;
-  text-decoration: none;
-}
