+++ src/pages/comparison/components/characteristics/characteristics.interface.ts
@@ -1,0 +1,5 @@
+import { PetsToCompareList } from '../../comparison.interface';
+
+export interface CharacteristicProps {
+  data: PetsToCompareList;
+}
