+++ src/pages/comparison/components/comparison-element/comparison-element.scss
@@ -1,0 +1,18 @@
+.item-compare {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 10px 0;
+  border-radius: 50px;
+  padding: 3px;
+
+  &:hover {
+    cursor: pointer;
+    background-color: rgb(243, 240, 240);
+  }
+}
+
+.link-style {
+  color: black;
+  text-decoration: none;
+}
