+++ src/pages/comparison/components/profile/profile.interface.ts
@@ -1,0 +1,4 @@
+export interface ProfileProps {
+  breed: string;
+  imgUrl: string;
+}
