+++ src/pages/comparison/components/profile/profile.scss
@@ -1,0 +1,17 @@
+.pet-block {
+  display: flex;
+  align-items: center;
+
+  .pet-image {
+    width: 35px;
+    height: 35px;
+    border-radius: 50px;
+    margin-right: 15px;
+  }
+
+  .pet-name {
+    font-size: 15px;
+    width: 150px;
+    line-height: 16px;
+  }
+}
