+++ src/pages/comparison/components/rate/rate.interface.ts
@@ -1,0 +1,3 @@
+export interface RateProps {
+  rate: number;
+}
