+++ src/pages/comparison/components/rate/rate.scss
