--- src/shared/logotype/logo.css
@@ -1,4 +1,0 @@
-.logo {
-    margin-left: -7px;
-    margin-bottom: -11px;
-}
