+++ src/store/root.reducer.ts
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 import { resultReducer } from './result/reducers/result.reducer';
 import { petReducer } from './pet/reducers/pet.reducer';
 import filterReducer from './filter/reducers/filter.reducer';
-import comparisonReducer from './reducers/comparison.reducer';
+import { comparisonReducer } from './comparison/reducers/comparison.reducer';
 
 export const rootReducer = combineReducers({
   filter: filterReducer,
