+++ src/store/root.saga.ts
@@ -1,7 +1,7 @@
 import createSagaMiddleware from 'redux-saga';
 import { all, fork } from 'redux-saga/effects';
 import { resultSaga } from './result/sagas/result.saga';
-import { comparisonSaga } from './sagas/comparison.saga';
+import { comparisonSaga } from './comparison/sagas/comparison.saga';
 import { petSaga } from './pet/sagas/pet.saga';
 
 export const saga = createSagaMiddleware();
